refactor(category): fix component name and delete unused components
