--- FinEdgeFrontend/src/Data/Categories/Categories.tsx
@@ -1,13 +1,0 @@
-import CreateCategory from './CreateCategory'
-import GetCategories from './GetCategories'
-
-const Categories = () => {
-  return (
-    <>
-      <CreateCategory />
-      <GetCategories />
-    </>
-  )
-}
-
-export default Categories
