+++ FinEdgeFrontend/src/Pages/Categories.tsx
